refactor(entity): Updating datatype for column Ratings from integer to double

+++ server/Domain/Entities/Media.cs
@@ -14,7 +14,7 @@ public class Media
 
     public string ImageURL { get; set; } = string.Empty;
 
-    public int Ratings { get; set; }
+    public double Ratings { get; set; }
 
     public MediaType Type { get; set; }
     public MediaStatus Status { get; set; }
